pyproject.toml
[project]
name = "cfdoti"
version = "0.0.1"
description = "An extension of doit tailored for running tasks in a local computeFarm"
authors = [
    {name = "Stephen Gaito", email = "stephen@perceptisys.co.uk"},
]
dependencies = [
    "pyyaml>=6.0",
    "doit>=0.36.0",
    "pkgconfig>=1.5.5",
]
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

[project.scripts]
cfdoit = "cfdoit:cli.cli"

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.dev-dependencies]
dev = [
    "flake8>=6.1.0",
    "mypy>=1.4.1",
    "types-PyYAML>=6.0.12.11",
]
